+++ samples/usage/simple-script/simple-script.ts
@@ -1,7 +1,7 @@
 #!/usr/bin/env ts-node
 
 import { createPromptbookExecutor, createPromptbookLibraryFromSources } from '@promptbook/core';
-import { JavascriptEvalExecutionTools } from '@promptbook/execute-javascript';
+import { JavascriptExecutionTools } from '@promptbook/execute-javascript';
 import { OpenAiExecutionTools } from '@promptbook/openai';
 import { assertsExecutionSuccessful, executionReportJsonToString } from '@promptbook/utils';
 import colors from 'colors';
@@ -36,7 +36,7 @@ async function main() {
             apiKey: process.env.OPENAI_API_KEY,
         }),
         script: [
-            new JavascriptEvalExecutionTools({
+            new JavascriptExecutionTools({
                 isVerbose: true,
             }),
         ],