+++ src/execution/plugins/script-execution-tools/custom-function-async.test.ts.test.ts
@@ -6,7 +6,7 @@ import { PromptbookString } from '../../../types/PromptbookString';
 import { createPromptbookExecutor } from '../../createPromptbookExecutor';
 import { MockedEchoLlmExecutionTools } from '../llm-execution-tools/mocked/MockedEchoLlmExecutionTools';
 import { CallbackInterfaceTools } from '../user-interface-execution-tools/callback/CallbackInterfaceTools';
-import { JavascriptEvalExecutionTools } from './javascript/JavascriptEvalExecutionTools';
+import { JavascriptExecutionTools } from './javascript/JavascriptExecutionTools';
 
 describe('createPromptbookExecutor + custom async function ', () => {
     const promptbook = promptbookStringToJson(
@@ -37,7 +37,7 @@ describe('createPromptbookExecutor + custom async function ', () => {
         tools: {
             llm: new MockedEchoLlmExecutionTools({ isVerbose: true }),
             script: [
-                new JavascriptEvalExecutionTools({
+                new JavascriptExecutionTools({
                     isVerbose: true,
 
                     // Note: [🕎]