+++ src/execution/plugins/script-execution-tools/custom-function-with-dependencies.test.ts
@@ -6,7 +6,7 @@ import { PromptbookString } from '../../../types/PromptbookString';
 import { createPromptbookExecutor } from '../../createPromptbookExecutor';
 import { MockedEchoLlmExecutionTools } from '../llm-execution-tools/mocked/MockedEchoLlmExecutionTools';
 import { CallbackInterfaceTools } from '../user-interface-execution-tools/callback/CallbackInterfaceTools';
-import { JavascriptEvalExecutionTools } from './javascript/JavascriptEvalExecutionTools';
+import { JavascriptExecutionTools } from './javascript/JavascriptExecutionTools';
 
 describe('createPromptbookExecutor + custom function with dependencies', () => {
     const promptbook = promptbookStringToJson(
@@ -38,7 +38,7 @@ describe('createPromptbookExecutor + custom function with dependencies', () => {
         tools: {
             llm: new MockedEchoLlmExecutionTools({ isVerbose: true }),
             script: [
-                new JavascriptEvalExecutionTools({
+                new JavascriptExecutionTools({
                     isVerbose: true,
 
                     // Note: [🕎]